Move namespace to be non-product editor specific
